+++ IL2-SR-Common/Network/ReleaseMetadata.cs
@@ -6,8 +6,8 @@ namespace Ciribob.IL2.SimpleRadio.Standalone.Common.Network
         public const string GithubRepository = "IL2-SimpleRadioStandalone";
         public const string GithubUserAgent = GithubUsername + "_" + GithubRepository;
 
-        public const string Version = "1.0.4.6";
-        public const string ReleaseTag = "1.0.4.6";
+        public const string Version = "1.0.4.7";
+        public const string ReleaseTag = "1.0.4.7-beta.1";
         public const string MinimumProtocolVersion = "1.0.0.0";
     }
 }